don't diff children for unrelated subtrees
only diff children for VNodes where only the properties are patched.
for nodes with a VPatch.VNODE patch, the children of the patch
VNode will be rendered as part of the patch-op for VPatch.VNODE and so
there is no need to diff the children.

another way to think about this is that if the 2 VNodes aren't similar
enough then one replaces the other and we can skip diffing the children
since the subtrees are assumed to be completely unrelated.